refactor: Refatora o Mission Controller para incluir o ResponseEntity nas rotas, customizando as respostas das requisições.

+++ src/main/java/com/diego/CadastroDeNinjas/Missions/MissionController.java
@@ -1,6 +1,8 @@
 package com.diego.CadastroDeNinjas.Missions;
 
 import com.diego.CadastroDeNinjas.Ninjas.NinjaModel;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -14,27 +16,43 @@ public class MissionController {
     }
 
     @PostMapping("/add")
-    public MissionDTO addMission(@RequestBody MissionDTO missionDTO){
-        return missionService.addMission(missionDTO);
+    public ResponseEntity<String> addMission(@RequestBody MissionDTO missionDTO){
+        MissionDTO mission = missionService.addMission(missionDTO);
+        return ResponseEntity.status(HttpStatus.CREATED).body("Missão de " + missionDTO.getName() + " criada com sucesso!");
     }
 
     @GetMapping("/list")
-    public List<MissionDTO> listMissions(){
-        return missionService.listMissions();
+    public ResponseEntity<List<MissionDTO>> listMissions(){
+        return ResponseEntity.ok().body(missionService.listMissions());
     }
 
     @GetMapping("/list/{id}")
-    public MissionDTO listMissionById(@PathVariable Long id){
-        return missionService.listMissionById(id);
+    public ResponseEntity<?> listMissionById(@PathVariable Long id){
+        if(missionService.listMissionById(id) != null){
+            MissionDTO mission = missionService.listMissionById(id);
+            return ResponseEntity.ok().body(mission);
+        }else{
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Esse id não existe.");
+        }
     }
 
     @PutMapping("/change/{id}")
-    public MissionDTO changeMission(@PathVariable Long id, @RequestBody MissionDTO missionDTO){
-        return missionService.changeMission(id, missionDTO);
+    public ResponseEntity<String> changeMission(@PathVariable Long id, @RequestBody MissionDTO missionDTO){
+        if(missionService.listMissionById(id) != null){
+            missionService.changeMission(id, missionDTO);
+            return ResponseEntity.ok().body("Missão com ID " + id + " alterada com sucesso!");
+        }else{
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Missão com ID " + id + " não existe.");
+        }
     }
 
     @DeleteMapping("/delete/{id}")
-    public void deleteMission(@PathVariable Long id){
-        missionService.deleteMission(id);
+    public ResponseEntity<String> deleteMission(@PathVariable Long id){
+        if(missionService.listMissionById(id) != null){
+            missionService.deleteMission(id);
+            return ResponseEntity.ok().body("Missão com ID " + id + " deletada com sucesso!");
+        }else{
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Missão com ID " + id + " não existe.");
+        }
     }
 }
